+++ src/components/WeeklyQuestion/Answer/AnswerList.tsx
@@ -1,7 +1,7 @@
 import { DailyAnswerContentInfo } from '@models/apis/QuestionType'
 import { TextBubble } from '@components/atoms/TextBubble'
 import { useEffect, useState } from 'react'
-import { UserApi } from '@apis/user/UserApi'
+import { UserApi } from '@apis/UserApi'
 import styled from '@emotion/styled'
 
 interface BubbleProfile {
