+++ src/components/WeeklyQuestion/Answer/DailyAnswerListContainer.tsx
@@ -6,7 +6,7 @@ import {
 } from '@models/apis/QuestionType'
 import styled from '@emotion/styled'
 import { useQuery } from '@tanstack/react-query'
-import { WeeklyApi } from '@apis/weekly/WeeklyApi'
+import { WeeklyApi } from '@apis/WeeklyApi'
 import { useEffect, useState } from 'react'
 import Dimmer from '@components/layouts/Dimmer'
 import { useOutsideClick } from '@libs/hooks/useOutsideClick'
