+++ src/components/WeeklyQuestion/DragDropFriend.tsx
@@ -4,7 +4,7 @@ import { Text } from '@components/atoms/Text'
 import { FlexBox } from '@components/layouts/FlexBox'
 import { palette } from '@styles/theme/palette'
 import { Draggable } from 'react-beautiful-dnd'
-import { FriendsApi } from '@apis/friends/FriendsApi'
+import { FriendsApi } from '@apis/FriendsApi'
 import OpenEye from '@assets/icons/OpenEye'
 import Ordering from '@assets/icons/Ordering'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
