+++ src/components/auth/RequireAuth.tsx
@@ -5,7 +5,7 @@ import { useRecoilState } from 'recoil'
 import { useEffect } from 'react'
 import { axiosApi } from '@apis/axios'
 import { Outlet, Navigate } from 'react-router-dom'
-import { UserApi } from '@apis/user/UserApi'
+import { UserApi } from '@apis/UserApi'
 import { IsOnboard } from '@libs/store/onboard'
 import Loading from '@components/atoms/Loading'
 import useSnackBar from '@libs/hooks/useSnackBar'
