+++ src/components/friends/AddFriendById.tsx
@@ -4,7 +4,7 @@ import { FlexBox } from '@components/layouts/FlexBox'
 import { SearchInput } from '@components/atoms/Input/SearchInput'
 import { Text } from '@components/atoms/Text'
 import { useSetFriendRecoilState } from '@libs/hooks/useSetFriendRecoilState'
-import { FriendsApi } from '@apis/friends/FriendsApi'
+import { FriendsApi } from '@apis/FriendsApi'
 import SearchedFriendList from './SearchedFriendList'
 import { authState } from '@libs/store/auth'
 import { useRecoilValue } from 'recoil'
