+++ src/components/friends/AllFriends.tsx
@@ -12,7 +12,7 @@ import ListIcon from '@assets/icons/ListIcon'
 import FriendsMenuIcon from '@assets/icons/FriendsMenuIcon'
 import { authState } from '@libs/store/auth'
 import { ProfileState, profileState } from '@libs/store/profile'
-import { FriendsApi } from '@apis/friends/FriendsApi'
+import { FriendsApi } from '@apis/FriendsApi'
 import useRecoilToggle from '@libs/hooks/useRecoilToggle'
 import { RoundButton } from '@components/atoms/RoundButton'
 import ShareIcon from '@assets/icons/ShareIcon'
