+++ src/components/friends/PresentRecommend.tsx
@@ -3,7 +3,7 @@ import { SelectedProps } from '@models/apis/UserType'
 import styled from '@emotion/styled'
 import { ItemFilter } from '@assets/icons/ItemFilter'
 import { Text } from '@components/atoms/Text'
-import { FriendsApi } from '@apis/friends/FriendsApi'
+import { FriendsApi } from '@apis/FriendsApi'
 import { GiftFilter } from '@components/atoms/GiftFilter'
 import SortItem from './Bottomsheet/SortItem'
 import { useRecoilState, useRecoilValue } from 'recoil'
