+++ src/components/funnel/OneAnswerStep.tsx
@@ -1,6 +1,6 @@
 import { TasteType } from '@models/apis/TasteType'
 import { useQuery } from '@tanstack/react-query'
-import { FavorApi } from '@apis/favor/FavorApi'
+import { FavorApi } from '@apis/FavorApi'
 import { FlexBox } from '@components/layouts/FlexBox'
 import { Text } from '@components/atoms/Text'
 import { RoundButton } from '@components/atoms/RoundButton'
