+++ src/components/funnel/SearchAnswerStep.tsx
@@ -9,7 +9,7 @@ import { ChangeEvent, useState, useEffect } from 'react'
 import { Text } from '@components/atoms/Text'
 import SquareButton from '@components/atoms/SquareButton'
 import { useRef } from 'react'
-import { FavorApi } from '@apis/favor/FavorApi'
+import { FavorApi } from '@apis/FavorApi'
 import { useQuery } from '@tanstack/react-query'
 import { useRecoilState } from 'recoil'
 import { answerState } from '@libs/store/question'
