+++ src/components/profile/ShareProfile/ShareProfileInfo.tsx
@@ -6,7 +6,7 @@ import { theme } from '@styles/theme'
 import useGetDate from '@libs/hooks/useGetDate'
 import { authState } from '@libs/store/auth'
 import { profileQueryKeys } from '@constants/queryKeys/profileQueryKeys'
-import { UserApi } from '@apis/user/UserApi'
+import { UserApi } from '@apis/UserApi'
 import { TasteBoxVariantType } from '@models/apis/TasteType'
 import TifySmallLogo from '@assets/icons/TifySmallLogo'
 import Svg from '@components/atoms/Svg'
