+++ src/pages/home/Redirect/KaKaoRedirect.tsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react'
 import { OauthCodeResponse } from '@models/apis/AuthType'
 import { useMutation } from '@tanstack/react-query'
 import { useLocation } from 'react-router-dom'
-import { AuthApi } from '@apis/auth/AuthApi'
+import { AuthApi } from '@apis/AuthApi'
 import useAuthMutate from '@libs/hooks/mutations/useAuthMutate'
 import Loading from '@components/atoms/Loading'
