+++ src/pages/onboarding/details/selectInfo/SelectFavor.tsx
@@ -9,7 +9,7 @@ import { BeautyFavor } from '@components/onboarding/BeautyFavor'
 import { FashionFavor } from '@components/onboarding/FashionFavor'
 import { HobbyFavor } from '@components/onboarding/HobbyFavor'
 import { useEffect, useState } from 'react'
-import { OnboardingApi } from '@apis/onboarding/OnboardingApi'
+import { OnboardingApi } from '@apis/OnboardingApi'
 import { authState } from '@libs/store/auth'
 import { useNavigate } from 'react-router-dom'
 import { parseFavorBox } from '@utils/parseFavorBox'
