+++ src/pages/profile/editfavorBox/EditFavorBox.tsx
@@ -8,7 +8,7 @@ import { useMutation } from '@tanstack/react-query'
 import { useRecoilValue } from 'recoil'
 import { authState } from '@libs/store/auth'
 import { parseFavorBox } from '@utils/parseFavorBox'
-import { UserApi } from '@apis/user/UserApi'
+import { UserApi } from '@apis/UserApi'
 import useSnackBar from '@libs/hooks/useSnackBar'
 import { useNavigate } from 'react-router-dom'
 import { TasteType } from '@models/apis/TasteType'
