+++ src/pages/weekly/AnswerDailyQuestion.tsx
@@ -6,7 +6,7 @@ import styled from '@emotion/styled'
 import { theme } from '@styles/theme'
 import { useRecoilState } from 'recoil'
 import { questionState } from '@libs/store/question'
-import { WeeklyApi } from '@apis/weekly/WeeklyApi'
+import { WeeklyApi } from '@apis/WeeklyApi'
 import { useNavigate } from 'react-router-dom'
 import { motion } from 'framer-motion'
 import { LongInput } from '@components/atoms/Input/LongInput'
