+++ src/pages/weekly/WeeklyMainQuestion.tsx
@@ -2,7 +2,7 @@ import { Spacing } from '@components/atoms/Spacing'
 import { Week } from '@components/atoms/DayWeek/WeekGroup/Week'
 import styled from '@emotion/styled'
 import { dateState } from '@libs/store/date'
-import { WeeklyApi } from '@apis/weekly/WeeklyApi'
+import { WeeklyApi } from '@apis/WeeklyApi'
 import { useRecoilState } from 'recoil'
 import DailyQuestionBox from '@components/WeeklyQuestion/DailyQuestionBox'
 import { questionState } from '@libs/store/question'
